-Move printList function back into index.js -Change index ordering start from 0 to 1.
