+++ src/adjustItems.js
@@ -10,7 +10,10 @@ function addTodo(arr) {
 function deleteTodo(arr, index) {
   const cont = document.getElementById(index);
   cont.remove();
-  arr.splice(index, 1);
+  arr.splice(index - 1, 1);
+  for (let i = 0; i < arr.length; i += 1) {
+    arr[i].index = i + 1;
+  }
   localStorage.clear();
   localStorage.setItem('todoArray', JSON.stringify(arr));
 }
